+++ plugins/Marketplace/resources/qml/Materials.qml
@@ -5,7 +5,8 @@ import Marketplace 1.0 as Marketplace
 
 Packages
 {
-    model: Marketplace.PackageList
+    pageTitle: catalog.i18nc("@header", "Install Materials")
+    model: Marketplace.RemotePackageList
     {
         packageTypeFilter: "material"
     }
